+++ crates/icydb-schema-tests/src/test/relation.rs
@@ -9,9 +9,9 @@ use crate::prelude::*;
     pk = "id",
     fields(
         field(ident = "id", value(item(prim = "Ulid")), default = "Ulid::generate"),
-        field(ident = "a_id", value(item(rel = "EntityA"))),
-        field(ident = "b_id", value(item(rel = "EntityB", prim = "Nat16"))),
-        field(ident = "c_id", value(item(rel = "EntityC", prim = "Principal"))),
+        field(ident = "a", value(item(rel = "EntityA"))),
+        field(ident = "b", value(item(rel = "EntityB", prim = "Nat16"))),
+        field(ident = "c", value(item(rel = "EntityC", prim = "Principal"))),
     )
 )]
 pub struct HasRelation;
@@ -25,7 +25,7 @@ pub struct HasRelation;
     pk = "id",
     fields(
         field(ident = "id", value(item(prim = "Ulid")), default = "Ulid::generate"),
-        field(ident = "a_ids", value(many, item(rel = "EntityA"))),
+        field(ident = "a", value(many, item(rel = "EntityA"))),
     )
 )]
 pub struct HasManyRelation;
@@ -83,7 +83,7 @@ pub struct RelationOwner;
     pk = "id",
     fields(
         field(ident = "id", value(item(prim = "Ulid")), default = "Ulid::generate"),
-        field(ident = "owner_id", value(item(rel = "RelationOwner"))),
+        field(ident = "owner", value(item(rel = "RelationOwner"))),
     )
 )]
 pub struct RelationOwned;
@@ -98,7 +98,7 @@ pub struct RelationOwned;
     pk = "id",
     fields(
         field(ident = "id", value(item(prim = "Ulid")), default = "Ulid::generate"),
-        field(ident = "owner_id", value(item(rel = "RelationOwner"))),
+        field(ident = "owner", value(item(rel = "RelationOwner"))),
     )
 )]
 pub struct CrossCanisterRelation;